Rename the API0 to APIv0 and try and do some more exception throwing in saveconfig.php
